Minimizing existing windows before running GUI tests
We don't need to see the GitHub runner console polluting the background
